Do not show warning on cancel of unsigned doc (#25)
* Do not show warning on cancel of unsigned doc

IB-4444

* Fix. Do not show warning on cancel of unsigned doc

IB-4444
a)  Take a document and add it to a container (either by right-click or through digidoc->sign_documents).
    Press Cancel, pop-up should appear to ask whether to keep or remove the new container

b)  Open an existing unsigned container (either by double-click or by using digidoc->View signed document content).
    Press cancel, the file should be closed and nothing should be asked from the user

c)  Open an existing unsigned container (either by double-click or by using digidoc->View signed document content).
    Add a new file to it. Press cancel, the file should be closed and nothing should be asked from the user,
    the newly added file should be stored in the container (so the next time I open it the file is still there)

d)  Open an existing unsigned container with more than one file in it.
    Remove one file from the container (by pressing the red X next to the file name).
    Press cancel, nothing should be asked from the user. The next time you open the container the file that was
    previously deleted should no longer be in the container.


* When cancel is pressed we always end up in the "Viewing" view

IB-4444
In File Manager->select unsigned container->Left mouse button click->Open With->Digidoc Client->
after Cancel is pressed always end up in the "Viewing" view. The behaviour would remain the same as before.

+++ client/MainWindow.cpp
@@ -50,6 +50,7 @@ MainWindow::MainWindow( QWidget *parent )
 	: QWidget( parent )
 	, cardsGroup( new QActionGroup( this ) )
 	, quitOnClose( false )
+	, warnOnUnsignedDocCancel( true )
 	, prevpage( Home )
 	, message( 0 )
 {
@@ -279,7 +280,10 @@ void MainWindow::buttonClicked( int button )
 			tr("Documents (%1%2)").arg( "*.bdoc *.ddoc *.asice *.sce")
 				.arg(qApp->confValue(Application::PDFUrl).toString().isEmpty() ? "" : " *.pdf") );
 		if( !file.isEmpty() && doc->open( file ) )
+		{
+			warnOnUnsignedDocCancel = false;
 			setCurrentPage( doc->signatures().isEmpty() ? Sign : View );
+		}
 		break;
 	}
 	case HomeCrypt:
@@ -302,6 +306,7 @@ void MainWindow::buttonClicked( int button )
 				{
 					if( doc->open( f.absoluteFilePath() ) )
 					{
+						warnOnUnsignedDocCancel = false;
 						setCurrentPage( doc->signatures().isEmpty() ? Sign : View );
 						enableSign();
 					}
@@ -312,12 +317,15 @@ void MainWindow::buttonClicked( int button )
 				else if( !addFile( f.absoluteFilePath() ) )
 					break;
 			}
+			warnOnUnsignedDocCancel = true;
 			params.clear();
 			if( !doc->isNull() )
 				setCurrentPage( Sign );
 		}
 		else
 		{
+			warnOnUnsignedDocCancel = true;
+
 			QStringList list = FileDialog::getOpenFileNames( this, tr("Select documents") );
 			Q_FOREACH( const QString &file, list )
 			{
@@ -353,22 +361,31 @@ void MainWindow::buttonClicked( int button )
 		}
 		if( signContentView->model()->rowCount() )
 		{
-			QMessageBox msgBox( QMessageBox::Question, tr("Save container"),
-				tr("You added %n file(s) to container, but these are not signed yet.\n"
-					"Should I keep unsigned documents or remove these?", "", signContentView->model()->rowCount()) );
-			msgBox.addButton( tr("Remove"), QMessageBox::ActionRole );
-			QPushButton *keep = msgBox.addButton( tr("Keep"), QMessageBox::ActionRole );
-			msgBox.exec();
+			if ( warnOnUnsignedDocCancel )
+			{
+				QMessageBox msgBox( QMessageBox::Question, tr("Save container"),
+					tr("You added %n file(s) to container, but these are not signed yet.\n"
+						"Should I keep unsigned documents or remove these?", "", signContentView->model()->rowCount()) );
+				msgBox.addButton( tr("Remove"), QMessageBox::ActionRole );
+				QPushButton *keep = msgBox.addButton( tr("Keep"), QMessageBox::ActionRole );
+				msgBox.exec();
 
-			if( msgBox.clickedButton() == keep )
+				if( msgBox.clickedButton() == keep )
+				{
+					save();
+					setCurrentPage( View );
+					break;
+				}
+
+				if( QFile::exists( doc->fileName() ) )
+					QFile::remove( doc->fileName() );
+			}
+			else
 			{
 				save();
 				setCurrentPage( View );
 				break;
 			}
-
-			if( QFile::exists( doc->fileName() ) )
-				QFile::remove( doc->fileName() );
 		}
 	case IntroBack:
 		if( prevpage == View )
@@ -378,6 +395,7 @@ void MainWindow::buttonClicked( int button )
 		}
 	case ViewClose:
 		doc->clear();
+		warnOnUnsignedDocCancel = true;
 		if( quitOnClose )
 			close();
 		setCurrentPage( Home );
